MOO-71 Move trend widget config into msfcore_dashboard_app.json
The config for the trend result widget was in it own file (msfcore_result_trend_app.json). This commit deletes this file and moves the config into msfcore_dashboard_app.json where all the dashboard widgets are configured in order to maintain consistency in the way are configuring the widgets.